meson: fix pkg-config file generation
We want the version in the pkgconfig file to be the same as the
project version, instead of setting it manually (which one can
forget easily). Also, since we now use both_libraries instead of
shared_library we need to add '1.0' to the filebase to get the
correct file name, since both_libraries doesn't set it the same
way shared_library does ('playerctl-1.0' instead of 'playerctl.pc').