cli: handle nonfile uris in open command
The cli command `playerctl open [uri]` would always convert the uri to a
file uri so the absolute path of the file could be passed to the player
from the playerctl cwd. Only do that conversion if the relative location
exists and otherwise just pass in the arg directly. This is required to
handle altnernative uri schemes, such as spotifies.

see #122