attempt handling invalid bus address settings
If the bus address is set to an invalid path, we will get the file not
found error from GIO when connecting to a bus. This might occur very
rarely in environments where there is no system bus (like my test
environment currently) or where the user has set the
DBUS_SESSION_BUS_ADDRESS environment variable to an incorrect value. I
don't believe the first case will ever occur for a normal user, and in
the second case, it is almost certainly user error. So we might just
settle on logging a warning and pretending the bus doesn't exist.